[fdf][rust] Make C doc comments legible to bindgen
Updates comments in the fdf header files so that they are legible to
bindgen as doc comments. No content has been changed, only changing //
comments to /// where appropriate.

Also changes the constants in the symbols library to be const statics
instead of macros, which bindgen has an easier time understanding as
well.

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1668014
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>

+++ sdk/lib/driver/component/rust/src/macros.rs
@@ -20,7 +20,7 @@ pub use fdf_sys::{
 #[doc(hidden)]
 pub const fn make_driver_registration<T: Driver>() -> DriverRegistration {
     DriverRegistration {
-        version: DRIVER_REGISTRATION_VERSION_1 as u64,
+        version: DRIVER_REGISTRATION_VERSION_1,
         v1: DriverRegistration_driver_registration_v1 {
             initialize: Some(DriverServer::<T>::initialize),
             destroy: Some(DriverServer::<T>::destroy),
